// src/types/developments.ts
// Generated by https://quicktype.io
//
// To change quicktype's target language, run command:
//
//   "Set quicktype target language"

export interface Development {
  name: string;
  copy: Copy;
  type: string;
  target: string;
  term: string;
  metadataURI: string;
  imageUrl: string;
  tokens: Token[];
  tokenPrice: string;
  maxSupply: string;
  maxMintAmount: string;
}

export interface Copy {
  en: En;
}

export interface En {
  name: string;
  location: string;
  rentability: string;
  description: string;
  term: string;
}

export interface Token {
  name: string;
  address: string;
  priceFactor: string;
}